requirements.txt
# MarketPulse v2.0 Requirements
# Personal AI-Powered Trading System

# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.3
numpy==1.24.3
asyncio-throttle==1.0.2

# AI Provider SDKs
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Technical Analysis
ta-lib==0.4.28  # Note: Requires separate TA-Lib C library installation
pandas-ta==0.3.14b0
yfinance==0.2.33

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.15.0  # For LSTM
prophet==1.1.5  # For swing trading predictions
joblib==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1  # Optional, for caching

# Web & Streaming
websocket-client==1.6.4
aiohttp==3.9.1
requests==2.31.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Broker Integration
kiteconnect==5.0.0  # Zerodha Kite Connect

# Alerts & Notifications
python-telegram-bot==20.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1

# Data Processing
scipy==1.11.4
statsmodels==0.14.0
numexpr==2.8.7
bottleneck==1.3.7

# Logging & Monitoring
loguru==0.7.2
rich==13.7.0

# Optional: Advanced Features
# alpaca-trade-api==3.0.2  # Alternative broker
# ccxt==4.1.56  # Crypto trading
# backtrader==1.9.78.123  # Advanced backtesting